sim:epiphany: Avoid overflow when parsing options
Change integer types to avoid overflow when parsing memory options

+++ sim/epiphany/sim-if.c
@@ -78,8 +78,8 @@ struct emesh_params {
 
   /* Extra */
   unsigned epiphany_add_ext_ram;
-  int ext_ram_base;
-  int ext_ram_size;
+  int64_t ext_ram_base;
+  int64_t ext_ram_size;
 
   unsigned initialized;
 };
